fix: switch rndChance to u32

+++ src/pokemon.c
@@ -6185,7 +6185,7 @@ u32 GetRandomPokerusDays(void)
 
 void RandomlyGivePartyPokerus(struct Pokemon *party)
 {
-    u16 rndChance = Random();
+    u32 rndChance = Random();
     u32 infectionChance = (GetNumOwnedBadges() * (P_BADGE_BOOST_POKERUS_CHANCE)) + POKERUS_INFECTION_CHANCE;
 
     if (rndChance < infectionChance)
